# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LSP UGM API Documentation",
        "description": "API Documentation for LSP UGM Application",
        "contact": {
            "name": "Administrator",
            "email": "admin@lsp-ugm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "LSP UGM API Server"
        }
    ],
    "paths": {
        "/asesmen/ak01/{id_asesi}": {
            "get": {
                "tags": [
                    "AK01"
                ],
                "summary": "Mendapatkan data formulir AK01 untuk asesi",
                "description": "Get AK01 data for an Asesi",
                "operationId": "10347f0941a6265a7454f9777ed2eeab",
                "parameters": [
                    {
                        "name": "id_asesi",
                        "in": "path",
                        "description": "ID asesi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data formulir AK01",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "general_info": {
                                                    "properties": {
                                                        "nama_asesi": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "nama_asesor": {
                                                            "type": "string",
                                                            "example": "Jane Smith"
                                                        },
                                                        "nama_tuk": {
                                                            "type": "string",
                                                            "example": "TUK JTTC"
                                                        },
                                                        "judul_skema": {
                                                            "type": "string",
                                                            "example": "Programmer"
                                                        },
                                                        "kode_skema": {
                                                            "type": "string",
                                                            "example": "SKM-001"
                                                        },
                                                        "pelaksanaan_asesmen_disepakati_mulai": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "2025-01-15"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "ak01": {
                                                    "properties": {
                                                        "hasil_yang_akan_dikumpulkan": {
                                                            "type": "string",
                                                            "example": "Hasil unjuk kerja, portofolio, dan knowledge test"
                                                        },
                                                        "waktu_tanda_tangan_asesi": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "2025-05-14",
                                                            "nullable": true
                                                        },
                                                        "waktu_tanda_tangan_asesor": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "2025-05-15",
                                                            "nullable": true
                                                        },
                                                        "tanda_tangan_asesor": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "signature/ttd_asesor.png"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "record_exists": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asesi tidak ditemukan atau belum memiliki rincian asesmen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asesi tidak ditemukan"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/asesmen/ak01/asesi/save": {
            "post": {
                "tags": [
                    "AK01"
                ],
                "summary": "Menyimpan tanda tangan Asesi pada formulir AK01",
                "description": "Create or update AK01 data for Asesi",
                "operationId": "51acc8918742b190d1809aff2b999c0e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_asesi",
                                    "id_asesor"
                                ],
                                "properties": {
                                    "id_asesi": {
                                        "type": "string",
                                        "example": "ASESI2025XXXXX"
                                    },
                                    "id_asesor": {
                                        "type": "string",
                                        "example": "ASESOR2025XXXXX"
                                    },
                                    "is_signing": {
                                        "description": "Set to true to sign the form",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data AK01 berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data AK01 berhasil ditandatangani oleh Asesi"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asesi tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asesi tidak ditemukan"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Asesi tidak terdaftar dengan asesor ini",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asesi tidak terdaftar dengan asesor ini"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 403
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/asesmen/ak01/asesor/save": {
            "post": {
                "tags": [
                    "AK01"
                ],
                "summary": "Menyimpan data formulir AK01 oleh Asesor",
                "description": "Create or update AK01 data for Asesor",
                "operationId": "2456a598da1c04a8e9611c380ddc968f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_asesi",
                                    "id_asesor",
                                    "hasil_yang_akan_dikumpulkan"
                                ],
                                "properties": {
                                    "id_asesi": {
                                        "type": "string",
                                        "example": "ASESI2025XXXXX"
                                    },
                                    "id_asesor": {
                                        "type": "string",
                                        "example": "ASESOR2025XXXXX"
                                    },
                                    "hasil_yang_akan_dikumpulkan": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "Hasil Observasi Langsung"
                                        }
                                    },
                                    "is_signing": {
                                        "description": "Set to true if the asesor is signing the form",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data AK01 berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data AK01 berhasil disimpan oleh Asesor"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/asesmen/mapa01/save": {
            "post": {
                "tags": [
                    "MAPA01"
                ],
                "summary": "Menyimpan data formulir MAPA01 oleh Asesor",
                "description": "Create or update MAPA01 data for Asesor",
                "operationId": "ecf75dc63b3c0074c78c08b6c8f828cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_asesi",
                                    "id_asesor",
                                    "lingkungan",
                                    "peluang_untuk_mengumpulkan_bukti",
                                    "hubungan_antara_standar_kompetensi",
                                    "pelaksana_asesmen",
                                    "pihak_yang_relevan_untuk_dikonfirmasi",
                                    "tolak_ukur_asesmen"
                                ],
                                "properties": {
                                    "id_asesi": {
                                        "type": "string",
                                        "example": "ASESI2025XXXXX"
                                    },
                                    "id_asesor": {
                                        "type": "string",
                                        "example": "ASESOR2025XXXXX"
                                    },
                                    "pendekatan_asesmen_asesi": {
                                        "type": "string",
                                        "example": "Pekerja Berpengalaman, dimana berasal dari industri/tempat kerja yang dalam operasionalnya mampu telusur dengan standar kompetensi"
                                    },
                                    "tujuan_asesmen": {
                                        "type": "string",
                                        "example": "Sertifikasi"
                                    },
                                    "lingkungan": {
                                        "type": "string",
                                        "example": "Tempat kerja nyata"
                                    },
                                    "peluang_untuk_mengumpulkan_bukti": {
                                        "type": "string",
                                        "example": "Tersedia"
                                    },
                                    "hubungan_antara_standar_kompetensi": {
                                        "type": "string",
                                        "example": "\tBukti untuk mendukung asesmen / RPL"
                                    },
                                    "pelaksana_asesmen": {
                                        "type": "string",
                                        "example": "\tOleh Lembaga Sertifikasi"
                                    },
                                    "pihak_yang_relevan_untuk_dikonfirmasi": {
                                        "type": "string",
                                        "example": "Manajer sertifikasi LSP"
                                    },
                                    "tolak_ukur_asesmen": {
                                        "type": "string",
                                        "example": "Kriteria asesmen dari kurikulum pelatihan"
                                    },
                                    "is_signing": {
                                        "description": "Set to true if the asesor is signing the form",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data MAPA01 berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data MAPA01 berhasil disimpan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asesi tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asesi tidak ditemukan"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Asesi tidak terdaftar dengan asesor ini",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asesi tidak terdaftar dengan asesor ini"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 403
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/asesor/progressAsesi/{id}": {
            "get": {
                "tags": [
                    "Asesi"
                ],
                "summary": "Mendapatkan data progres asesmen asesi",
                "description": "Get asesi's progress asesmen",
                "operationId": "87ee5a5b03dd7ee23b2e0d032880eebe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID asesi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Progress Asesmen ditemukan"
                                        },
                                        "data": {
                                            "properties": {
                                                "progress_asesmen": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Progress Asesmen tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "API_KEY",
                "in": "header"
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "API_KEY",
                "in": "header"
            }
        }
    },
...
}